improve rendering
git-svn-id: https://osmand.googlecode.com/svn/trunk@493 e29c36b1-1cfa-d876-8d93-3434fc2bb7b8

+++ DataExtractionOSM/src/net/osmand/data/index/DataIndexWriter.java
@@ -457,10 +457,10 @@ public class DataIndexWriter {
 			addBatch(statements, mapStat);
 
 			mapWayLocationsStat.setLong(1, id);
-			mapWayLocationsStat.setDouble(2, minLon);
-			mapWayLocationsStat.setDouble(3, maxLon);
-			mapWayLocationsStat.setDouble(4, minLat);
-			mapWayLocationsStat.setDouble(5, maxLat);
+			mapWayLocationsStat.setFloat(2, (float) minLon);
+			mapWayLocationsStat.setFloat(3, (float) maxLon);
+			mapWayLocationsStat.setFloat(4, (float) minLat);
+			mapWayLocationsStat.setFloat(5, (float) maxLat);
 			addBatch(statements, mapWayLocationsStat);
 		}
 	}
